requirements.txt
aiofiles==23.1.0
aioredis==2.0.1
aiosqlite==0.17.0
anyio==3.6.2
async-timeout==4.0.2
Babel==2.12.1
bcrypt==4.0.1
certifi==2022.12.7
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.95.0
flake8==6.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
mypy==1.1.1
mypy-extensions==1.0.0
orjson==3.8.7
pendulum==2.1.2
pycodestyle==2.10.0
pydantic==1.10.6
pyflakes==3.0.1
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.7.1
pytzdata==2020.1
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqladmin==0.10.0
SQLAlchemy==2.0.7
starlette==0.26.1
tomli==2.0.1
tortoise-orm==0.19.3
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
WTForms==3.0.1
